Backport to branch(3.13) : Fix to avoid deleting non ScalarDB tables when dropping namespaces (#3100)

+++ core/src/main/java/com/scalar/db/storage/cosmos/CosmosAdmin.java
@@ -314,7 +314,15 @@ public class CosmosAdmin implements DistributedStorageAdmin {
   @Override
   public void dropNamespace(String namespace) throws ExecutionException {
     try {
+      Set<String> remainingTables = getRawTableNames(namespace);
+      if (!remainingTables.isEmpty()) {
+        throw new IllegalArgumentException(
+            CoreError.NAMESPACE_WITH_NON_SCALARDB_TABLES_CANNOT_BE_DROPPED.buildMessage(
+                namespace, remainingTables));
+      }
       client.getDatabase(namespace).delete();
+    } catch (IllegalArgumentException e) {
+      throw e;
     } catch (RuntimeException e) {
       throw new ExecutionException("Deleting the database failed", e);
     }
@@ -613,5 +621,11 @@ public class CosmosAdmin implements DistributedStorageAdmin {
       throw e;
     }
     return true;
+  }
+
+  private Set<String> getRawTableNames(String namespace) {
+    return client.getDatabase(namespace).readAllContainers().stream()
+        .map(CosmosContainerProperties::getId)
+        .collect(Collectors.toSet());
   }
 }
